requirements.txt
bert-extractive-summarizer==0.4.0
Flask
Flask-RESTful
Flask-SQLAlchemy
uwsgi